requirements.txt
cdasgiref==3.4.1
alembic==1.7.1
blueprint==3.4.2
backcall==0.2.0
cachelib==0.3.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
charset-normalizer==2.0.5
dnspython==2.1.0
email-validator==1.1.2
decorator==5.0.9
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-OAuthlib==0.9.6
Flask-Session==0.4.0
Flask-WTF==0.15.1
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
greenlet==1.1.1
idna==3.2
importlib-metadata==4.6.4
ipdb==0.13.9
ipython==7.26.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mysql-connector-python==8.0.26
oauthlib==2.1.0
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.19
protobuf==3.17.3
pyChart.JS==0.3.0
pychartjs==0.9.2
pycparser==2.20
Pygments==2.10.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.6
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
xero-python==1.9.0
zipp==3.5
